@@ -25,12 +25,12 @@ class HttpRequest {
 	private $tmp;
 
 	private $maxRedirects = 10;
-	private $headers = array();
+	private $headers = [];
 	private $endpointParts;
-	private $cookies = array();
+	private $cookies = [];
 	private $cookiesFollowRedirects = false;
 	/** @var array|string */
-	private $body = array();
+	private $body = [];
 	private $lastRequestTime = null;
 
 	/**
@@ -86,19 +86,19 @@ class HttpRequest {
 	 * Retrieve a url param by name
 	 *
 	 * @param string $param The name of the param.
-	 * @return string|array|null Null on failure.
+	 * @return array|string|null Null on failure.
 	 */
 	public function getUrlParam( $param ) {
 		$params = $this->getUrlParams();
 
-		return isset($params[$param]) ? $params[$param] : null;
+		return $params[$param] ?? null;
 	}
 
 	/**
 	 * Set a url param by name.
 	 *
 	 * @param string                 $param The name of the param.
-	 * @param string|int|float|array $value
+	 * @param array|float|int|string $value
 	 */
 	public function setUrlParam( $param, $value ) {
 		$params         = $this->getUrlParams();
@@ -201,7 +201,6 @@ class HttpRequest {
 	 * @deprecated Use setMethod and setFormValue instead
 	 *
 	 * @param string $key
-	 * @param mixed  $value
 	 */
 	public function setPost( $key, $value ) {
 		$this->method = self::POST;
@@ -214,7 +213,6 @@ class HttpRequest {
 	 * Note that if there is a non-form body set, this will replace it.
 	 *
 	 * @param string $key
-	 * @param mixed  $value
 	 */
 	public function setFormValue( $key, $value ) {
 		if( !is_array($this->body) ) {
@@ -250,8 +248,6 @@ class HttpRequest {
 	 * Note that this has the side effect of changing the HTTP Method to POST
 	 *
 	 * @deprecated Use setBody instead
-	 *
-	 * @param array $post
 	 */
 	public function setPostData( array $post ) {
 		$this->method = self::POST;
@@ -307,17 +303,16 @@ class HttpRequest {
 	}
 
 	/**
-	 * @param array $parsed_url
 	 * @return string
 	 */
 	private function composeUrl( array $parsed_url ) {
 		$scheme   = isset($parsed_url['scheme']) ? $parsed_url['scheme'] . '://' : '';
-		$host     = isset($parsed_url['host']) ? $parsed_url['host'] : '';
+		$host     = $parsed_url['host'] ?? '';
 		$port     = isset($parsed_url['port']) ? ':' . $parsed_url['port'] : '';
-		$user     = isset($parsed_url['user']) ? $parsed_url['user'] : '';
+		$user     = $parsed_url['user'] ?? '';
 		$pass     = isset($parsed_url['pass']) ? ':' . $parsed_url['pass'] : '';
 		$pass     = ($user || $pass) ? "{$pass}@" : '';
-		$path     = isset($parsed_url['path']) ? $parsed_url['path'] : '';
+		$path     = $parsed_url['path'] ?? '';
 		$query    = isset($parsed_url['query']) ? '?' . $parsed_url['query'] : '';
 		$fragment = isset($parsed_url['fragment']) ? '#' . $parsed_url['fragment'] : '';
 
@@ -427,7 +422,7 @@ class HttpRequest {
 	 * @return array
 	 */
 	private function getFlatHeaders() {
-		$output = array();
+		$output = [];
 		foreach( $this->getHeaders() as $key => $value ) {
 			$output[] = "$key: $value";
 		}
@@ -435,9 +430,6 @@ class HttpRequest {
 		return $output;
 	}
 
-	/**
-	 * @return mixed
-	 */
 	public function getCurlInfo() {
 		return $this->curlInfo;
 	}
@@ -445,7 +437,7 @@ class HttpRequest {
 	/**
 	 * Get the time the last request took in seconds a float
 	 *
-	 * @return null|float null if there is no last request
+	 * @return float|null null if there is no last request
 	 */
 	public function getLastRequestTime() {
 		return $this->lastRequestTime;
