@@ -12,7 +12,7 @@ if( isset($argv[1]) ) {
 }else{
 	$fpath = $buildRoot . "/boomerang.phar";
 }
- 
+
 $phar = new Phar($fpath, 
 	FilesystemIterator::CURRENT_AS_FILEINFO |     	FilesystemIterator::KEY_AS_FILENAME, "myapp.phar");
 
