@@ -7,6 +7,6 @@ $response = $req->makeRequest();
 
 $valid = new HttpResponseValidator($response);
 $valid->expectStatus(200)
-	  ->expectHeader('Content-Type', 'application/json');
+	->expectHeader('Content-Type', 'application/json');
 
 Boomerang::addValidator($valid);
