@@ -12,18 +12,13 @@ class PassingExpectationResult extends PassingResult {
 	protected $actual;
 
 	/**
-	 * @param ValidatorInterface $validator
-	 * @param null|string        $message
-	 * @param mixed              $actual
+	 * @param string|null        $message
 	 */
 	public function __construct( ValidatorInterface $validator, $message = null, $actual = null ) {
 		parent::__construct($validator, $message);
 		$this->actual = $actual;
 	}
 
-	/**
-	 * @return mixed
-	 */
 	public function getActual() {
 		return $this->actual;
 	}
