@@ -20,9 +20,6 @@ class HttpResponseValidator extends AbstractValidator {
 	 */
 	private $response;
 
-	/**
-	 * @param HttpResponseInterface $response
-	 */
 	public function __construct( HttpResponseInterface $response ) {
 		$this->response = $response;
 	}
@@ -59,7 +56,7 @@ class HttpResponseValidator extends AbstractValidator {
 	 *
 	 * @param string   $key The header field name to verify.
 	 * @param string   $value The expected value.
-	 * @param null|int $hop The zero indexed redirect hop. Defaults to the final hop.
+	 * @param int|null $hop The zero indexed redirect hop. Defaults to the final hop.
 	 * @return $this
 	 */
 	public function expectHeader( $key, $value, $hop = null ) {
@@ -81,7 +78,7 @@ class HttpResponseValidator extends AbstractValidator {
 	 *
 	 * @param string   $key The header field name to verify.
 	 * @param string   $value The expected containing value.
-	 * @param null|int $hop The zero indexed redirect hop. Defaults to the final hop.
+	 * @param int|null $hop The zero indexed redirect hop. Defaults to the final hop.
 	 * @return $this
 	 */
 	public function expectHeaderContains( $key, $value, $hop = null ) {
