@@ -38,11 +38,10 @@ class TestRunner {
 			$dir   = new \RecursiveDirectoryIterator($path);
 			$ite   = new \RecursiveIteratorIterator($dir);
 
-			return new \RegexIterator($ite, "/Spec\.php$/");
+			return new \RegexIterator($ite, "/Spec\\.php$/");
 		}
-
 		if( is_readable($path) ) {
-			return new \ArrayIterator(array( $path ));
+			return new \ArrayIterator([ $path ]);
 		}
 
 		throw new CliRuntimeException("Cannot find file \"$path\"");
