@@ -25,9 +25,6 @@ class AllEx extends StructureEx {
 	 */
 	protected $structures;
 
-	/**
-	 * @param TypeExpectationInterface|callable|mixed $structure,... One or more structure definitions to match
-	 */
 	public function __construct( $structure ) {
 		$this->structures = func_get_args();
 	}
@@ -36,7 +33,7 @@ class AllEx extends StructureEx {
 		$pass = true;
 
 		foreach( $this->structures as $struct ) {
-			list($passing, $expectations) = $this->__validate($data, $struct);
+			[$passing, $expectations] = $this->__validate($data, $struct);
 			$this->addExpectationResults($expectations);
 			$pass = $pass && $passing;
 		}
