@@ -4,14 +4,15 @@ namespace Boomerang\ExpectationResults\Test;
 
 use Boomerang\ExpectationResults\FailingResult;
 use Boomerang\Interfaces\ValidatorInterface;
+use PHPUnit\Framework\TestCase;
 
-class FailingResultTest extends \PHPUnit_Framework_TestCase {
+class FailingResultTest extends TestCase {
 
-	function testGetFail() {
+	public function testGetFail() {
 		/**
 		 * @var $mockValidator ValidatorInterface
 		 */
-		$mockValidator = $this->getMock('Boomerang\\Interfaces\\ValidatorInterface');
+		$mockValidator = $this->getMockBuilder(ValidatorInterface::class)->getMock();
 		$result        = new FailingResult($mockValidator);
 		$this->assertTrue($result->getFail());
 	}
