@@ -3,8 +3,9 @@
 namespace Boomerang\TypeExpectations\Test;
 
 use Boomerang\TypeExpectations\IntEx;
+use PHPUnit\Framework\TestCase;
 
-class IntExTest extends \PHPUnit_Framework_TestCase {
+class IntExTest extends TestCase {
 
 	public function testBasicMatching() {
 
@@ -15,7 +16,7 @@ class IntExTest extends \PHPUnit_Framework_TestCase {
 		$this->assertEquals(true, $x->match(-1000));
 
 		$this->assertEquals(false, $x->match(1.1));
-		$this->assertEquals(false, $x->match(array( 1 )));
+		$this->assertEquals(false, $x->match([ 1 ]));
 		$this->assertEquals(false, $x->match("1"));
 		$this->assertEquals(false, $x->match(true));
 
@@ -33,7 +34,6 @@ class IntExTest extends \PHPUnit_Framework_TestCase {
 		$this->assertEquals(true, $x->match(10.0));
 		$this->assertEquals(true, $x->match(-10.0));
 
-
 		$this->assertEquals(false, $x->match(9.9));
 		$this->assertEquals(false, $x->match(-9.9));
 		$this->assertEquals(false, $x->match(11));
@@ -42,7 +42,7 @@ class IntExTest extends \PHPUnit_Framework_TestCase {
 		$this->assertEquals(false, $x->match(-1000));
 
 //		$this->assertEquals(false, $x->match(9.9999999999999999999999));  this one is troublesome
-		$this->assertEquals(false, $x->match(array( 1 )));
+		$this->assertEquals(false, $x->match([ 1 ]));
 		$this->assertEquals(false, $x->match("1"));
 		$this->assertEquals(false, $x->match("-1"));
 		$this->assertEquals(false, $x->match(false));
